remove inflection grouping into language model
