+++ src/language_model.js
@@ -1,6 +1,8 @@
 import * as Constants from './constants.js'
 import Feature from './feature.js'
 import FeatureType from './feature_type.js'
+import InflectionGroupingKey from './inflection_grouping_key.js'
+import InflectionGroup from './inflection_group.js'
 
 /**
  * @class  LanguageModel is the base class for language-specific behavior
@@ -203,6 +205,131 @@ class LanguageModel {
    */
   static isLanguageCode (language) {
     return !LanguageModel.isLanguageID(language)
+  }
+
+  /**
+   * Groups a set of inflections according to a language-specific display paradigm
+   * The default groups according to the following logic:
+   *   1. groups of groups with unique stem, prefix, suffix, part of speech dialect and comparison
+   *     2. groups of those groups with unique
+   *          number, if it's an inflection with a grammatical case
+   *          tense, if it's an inflection with tense but no case (i.e. a verb)
+   *          verbs without tense or case
+   *          adverbs
+   *          everything else
+   *       3. groups of those groups with unique voice and tense
+   *         4. groups of inflections with unique gender, person, mood, and sort
+   */
+  groupInflectionsForDisplay (inflections) {
+    let grouped = new Map()
+
+    // group inflections by part of speech
+    for (let infl of inflections) {
+      let groupingKey = new InflectionGroupingKey(infl,
+        [Feature.types.part, Feature.types.dialect, Feature.types.comparison],
+        { prefix: infl.prefix,
+          suffix: infl.suffix,
+          stem: infl.stem
+        }
+        )
+      let groupingKeyStr = groupingKey.toString()
+      if (grouped.has(groupingKeyStr)) {
+        grouped.get(groupingKeyStr).append(infl)
+      } else {
+        grouped.set(groupingKeyStr, new InflectionGroup(groupingKey, [infl]))
+      }
+    }
+
+    // iterate through each group key to group the inflections in that group
+    for (let kv of grouped) {
+      let inflgrp = new Map()
+      for (let infl of kv[1].inflections) {
+        let keyprop
+        let isCaseInflectionSet = false
+        if (infl[Feature.types.grmCase]) {
+          // grouping on number if case is defined
+          keyprop = Feature.types.number
+          isCaseInflectionSet = true
+        } else if (infl[Feature.types.tense]) {
+          // grouping on tense if tense is defined but not case
+          keyprop = Feature.types.tense
+        } else if (infl[Feature.types.part] === Constants.POFS_VERB) {
+          // grouping on no case or tense but a verb
+          keyprop = Feature.types.part
+        } else if (infl[Feature.types.part] === Constants.POFS_ADVERB) {
+          keyprop = Feature.types.part
+          // grouping on adverbs without case or tense
+        } else {
+          keyprop = 'misc'
+          // grouping on adverbs without case or tense
+          // everything else
+        }
+        let groupingKey = new InflectionGroupingKey(infl, [keyprop], {isCaseInflectionSet: isCaseInflectionSet})
+        let groupingKeyStr = groupingKey.toString()
+        if (inflgrp.has(groupingKeyStr)) {
+          inflgrp.get(groupingKeyStr).append(infl)
+        } else {
+          inflgrp.set(groupingKeyStr, new InflectionGroup(groupingKey, [infl]))
+        }
+      }
+      // inflgrp is now a map of groups of inflections grouped by
+      //  inflections with number
+      //  inflections without number but with tense
+      //  inflections of verbs without tense
+      //  inflections of adverbs
+      //  everything else
+      // iterate through each inflection group key to group the inflections in that group by tense and voice
+      for (let kv of inflgrp) {
+        let nextGroup = new Map()
+        let sortOrder = new Map()
+        for (let infl of kv[1].inflections) {
+          let sortkey = infl[Feature.types.grmCase] ? Math.max(infl[Feature.types.grmCase].map((f) => { return f.sortOrder })) : 1
+          let groupingKey = new InflectionGroupingKey(infl, [Feature.types.tense, Feature.types.voice])
+          let groupingKeyStr = groupingKey.toString()
+          if (nextGroup.has(groupingKeyStr)) {
+            nextGroup.get(groupingKeyStr).append(infl)
+          } else {
+            nextGroup.set(groupingKeyStr, new InflectionGroup(groupingKey, [infl], sortkey))
+            sortOrder.set(groupingKeyStr, sortkey)
+          }
+        }
+        kv[1].inflections = []
+        let sortedKeys = Array.from(nextGroup.keys()).sort(
+          (a, b) => {
+            let orderA = sortOrder.get(a)
+            let orderB = sortOrder.get(b)
+            return orderA > orderB ? -1 : orderB > orderA ? 1 : 0
+          }
+        )
+        for (let groupkey of sortedKeys) {
+          kv[1].inflections.push(nextGroup.get(groupkey))
+        }
+      }
+
+      // inflgrp is now a Map of groups of groups of inflections
+
+      for (let kv of inflgrp) {
+        let groups = kv[1]
+        for (let group of groups.inflections) {
+          let nextGroup = new Map()
+          for (let infl of group.inflections) {
+            // set key is case comp gend pers mood sort
+            let groupingKey = new InflectionGroupingKey(infl,
+              [Feature.types.grmCase, Feature.types.comparison, Feature.types.gender, Feature.types.number, Feature.types.person,
+                Feature.types.tense, Feature.types.mood, Feature.types.sort, Feature.types.voice])
+            let groupingKeyStr = groupingKey.toString()
+            if (nextGroup.has(groupingKeyStr)) {
+              nextGroup.get(groupingKeyStr).append(infl)
+            } else {
+              nextGroup.set(groupingKeyStr, new InflectionGroup(groupingKey, [infl]))
+            }
+          }
+          group.inflections = Array.from(nextGroup.values()) // now a group of inflection groups
+        }
+      }
+      kv[1].inflections = Array.from(inflgrp.values())
+    }
+    return Array.from(grouped.values())
   }
 }
